// src/api/models/account-body-req.ts
/* tslint:disable */
/* eslint-disable */
/**
 * S2I final
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.8
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AccountBodyReq
 */
export interface AccountBodyReq {
    /**
     * 
     * @type {string}
     * @memberof AccountBodyReq
     */
    'username'?: string;
    /**
     * 
     * @type {string}
     * @memberof AccountBodyReq
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof AccountBodyReq
     */
    'surname'?: string;
    /**
     * 
     * @type {string}
     * @memberof AccountBodyReq
     */
    'email'?: string;
    /**
     * 
     * @type {string}
     * @memberof AccountBodyReq
     */
    'password'?: string;
}